Rankings: make updateWith() thread-safe
This fixes a ConcurrentModificationException that could be triggered when
using log4j to log Rankings#toString() from within a bolt's execute()
method.  Apparently a separate thread is used by log4j to perform the
logging, whose execution path could interweave with the bolt's own
thread.

Even though this is not necessarily a deficiency of the Rankings class
(which does not claim to be thread-safe), we still decided to make
Rankings#updateWith() thread-safe so that users do not run into errors
when using log4j & Co. for logging purposes.  In other words, we decided
to add this fix for the convenience of our users.