Update nimbus.clj
Adding check to disallow topology submission if the topology name is blank. In one of the scenarios, we had pushed a topology with an empty name, and it showed up on the ui with only the uid. Also since it did not have any name we could not go into "Topology Details" to kill it, and hence had to clean up the state on the disk so that Nimbus kills the topology.