[STORM-1175] State store for windowing operations
Added IStatefulBolt abstraction that can be implemented by bolts which requires
its state to be checkpointed periodically. State implementations based on
key-value mapping store are added. There is a default in-memory based implementation
and optional redis based implementation that provides state persistence. An internal
CheckpointSpout periodically emits checkpoint tuples which flows through the
topology DAG to take a consistent snapshot of the state across all components.

There is still pending work to capture the evaluated/expired state of the tuples
in the Window and use it to prune duplicate window evaluations during restart. This can
be built on top of the stateful bolt abstractions and will be done as part of STORM-1176